backend/requirements.txt
alembic==1.8.1
Babel==2.10.3
blinker==1.5
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
Flask==2.2.1
Flask-Babel==2.0.0
Flask-Login==0.6.2
Flask-Mail==0.9.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
SQLAlchemy==1.4.39
urllib3==1.26.12
Werkzeug==2.2.1
WTForms==3.0.1
zipp==3.8.1
